fix(aapd-1698): use token to lookup user's draft appeals
