--- packages/appeals-service-api/src/routes/v2/users/_id/appeals/index.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const { getUserAppeals } = require('./controller');
-const asyncHandler = require('@pins/common/src/middleware/async-handler');
-const router = express.Router({ mergeParams: true });
-
-router.get('/', asyncHandler(getUserAppeals));
-
-module.exports = { router };
