+++ packages/forms-web-app/src/controllers/appeals/your-appeals/decided-appeals.js
@@ -1,15 +1,12 @@
 const { mapToAppellantDashboardDisplayData } = require('../../../lib/dashboard-functions');
 const { VIEW } = require('../../../lib/views');
 const logger = require('../../../lib/logger');
-const { apiClient } = require('../../../lib/appeals-api-client');
 const { sortByDateFieldDesc } = require('@pins/common/src/lib/appeal-sorting');
 
 exports.get = async (req, res) => {
-	const { email } = req.session;
 	let viewContext = {};
 	try {
-		const user = await apiClient.getUserByEmailV2(email);
-		const appeals = await apiClient.getUserAppealsById(user.id);
+		const appeals = await req.appealsApiClient.getUserAppeals();
 		if (appeals?.length > 0) {
 			const decidedAppeals = appeals
 				.map(mapToAppellantDashboardDisplayData)
